src/main/java/com/ecoinv/inventorymanagement/InventoryManager.java
package com.ecoinv.inventorymanagement;

import java.util.List;

public class InventoryManager {
    // Constructor
    public InventoryManager() {
        // Initialization code here (e.g., setting up API clients, database connections)
    }

    // Method to analyze sales data
    public void analyzeSalesData() {
        // Fetch sales data from a database or an external API
        // Process the data to identify sales patterns and trends
        System.out.println("Analyzing sales data...");
        // This could involve complex data analysis and machine learning predictions
        // Consider using a library or service for ML predictions
    }

    // Method to adjust inventory levels based on analyzed data
    public void adjustInventoryLevels() {
        // Based on the analysis, determine which products need more stock and which have excess
        System.out.println("Adjusting inventory levels...");
        // Adjust inventory records in the database or through an inventory management API
        // This step might involve setting thresholds for reordering or reducing stock
    }

    // Method for automated reordering with supplier systems
    public void automateReordering() {
        // Identify items that need reordering
        System.out.println("Automating reordering process...");
        // Send orders to suppliers via their APIs or by generating purchase orders
        // Consider implementing a confirmation and tracking system for orders
    }

    // Additional methods can be defined here for tasks like:
    // - Integrating with market APIs to fetch trend data
    // - Communicating with machine learning services for demand forecasting
}
