src/test/java/com/ecoinv/inventorymanagement/InventoryManagerTest.java
package com.ecoinv.inventorymanagement;
import org.junit.After;
import org.junit.Before;
import org.junit.Test;
import static org.junit.Assert.*;

public class InventoryManagerTest {

    private InventoryManager inventoryManager;

    // Setup before each test
    @Before
    public void setUp() {
        inventoryManager = new InventoryManager();
        // Initialize your InventoryManager with any required data
    }

    // Test case for analyzing sales data
    @Test
    public void testAnalyzeSalesData() {
        // Call the method you want to test
        inventoryManager.analyzeSalesData();
        // Assert the expected outcomes using assertTrue, assertEquals, etc.
        // For example:
        // assertEquals("Expected result", inventoryManager.getSomeResult());
    }

    // Test case for adjusting inventory levels
    @Test
    public void testAdjustInventoryLevels() {
        // Setup specific conditions for the test
        // Call the method to test
        inventoryManager.adjustInventoryLevels();
        // Assert the expected outcomes
    }

    // Test case for automated reordering
    @Test
    public void testAutomateReordering() {
        // Setup specific conditions for the test
        // Call the method to test
        inventoryManager.automateReordering();
        // Assert the expected outcomes
    }

    // Cleanup after each test
    @After
    public void tearDown() {
        // Clean up resources, if necessary
    }
}
